cleanup size_t/uint mismatch to get less compiler warnings on 64bit Windows cross compiling.
[git-p4: depot-paths = "//depot/qt/main/": change = 44528]

+++ src/tools/qcstring.h
@@ -55,7 +55,7 @@ Q_EXPORT void *qmemmove( void *dst, const void *src, uint len );
 Q_EXPORT char *qstrdup( const char * );
 
 Q_EXPORT inline uint qstrlen( const char *str )
-{ return str ? strlen(str) : 0; }
+{ return str ? (uint)strlen(str) : 0; }
 
 Q_EXPORT inline char *qstrcpy( char *dst, const char *src )
 { return src ? strcpy(dst, src) : 0; }
